Fix and move tests
Fix some broken tests and move them into the appropriate directories.
