+++ test/unit/sync_checker/request_queue_test.rb
@@ -4,6 +4,7 @@ require 'plek'
 require 'minitest/autorun'
 require 'mocha/setup'
 require_relative '../../../lib/sync_checker/request_queue'
+require_relative '../../../test/test_helper'
 
 class SyncChecker::RequestQueueTest < Minitest::Test
   def setup
@@ -24,7 +25,7 @@ class SyncChecker::RequestQueueTest < Minitest::Test
     mutex = Mutex.new
     queued_request = SyncChecker::RequestQueue.new(document_check, result_set, mutex)
     response = Typhoeus::Response.new(code: 200, body: "{'content_id', 'booyah'}")
-    Typhoeus.stub("http://draft-content-store.dev.gov.uk/content/one").and_return(response)
+    Typhoeus.stub("https://draft-content-store.test.alphagov.co.uk/content/one").and_return(response)
 
     document_check.expects(:check_draft).with(response, :en)
     queued_request.requests.map(&:run)
@@ -40,7 +41,7 @@ class SyncChecker::RequestQueueTest < Minitest::Test
     mutex = Mutex.new
     queued_request = SyncChecker::RequestQueue.new(document_check, result_set, mutex)
     response = Typhoeus::Response.new(code: 200, body: "{'content_id', 'booyah'}")
-    Typhoeus.stub("http://content-store.dev.gov.uk/content/one").and_return(response)
+    Typhoeus.stub("https://content-store.test.alphagov.co.uk/content/one").and_return(response)
 
     document_check.expects(:check_live).with(response, :en)
     queued_request.requests.map(&:run)
@@ -56,7 +57,7 @@ class SyncChecker::RequestQueueTest < Minitest::Test
     mutex = Mutex.new
     queued_request = SyncChecker::RequestQueue.new(document_check, result_set, mutex)
     response = Typhoeus::Response.new(code: 200, body: "{'content_id', 'booyah'}")
-    Typhoeus.stub("http://content-store.dev.gov.uk/content/one").and_return(response)
+    Typhoeus.stub("https://content-store.test.alphagov.co.uk/content/one").and_return(response)
 
     document_check.expects(:check_live).with(response, :en).returns(check_result = stub)
     queued_request.requests.map(&:run)
@@ -74,7 +75,7 @@ class SyncChecker::RequestQueueTest < Minitest::Test
     mutex = Mutex.new
     queued_request = SyncChecker::RequestQueue.new(document_check, result_set, mutex)
     response = Typhoeus::Response.new(code: 200, body: "{'content_id', 'booyah'}")
-    Typhoeus.stub("http://draft-content-store.dev.gov.uk/content/one").and_return(response)
+    Typhoeus.stub("https://draft-content-store.test.alphagov.co.uk/content/one").and_return(response)
 
     document_check.expects(:check_draft).with(response, :en).returns(check_result = stub)
     queued_request.requests.map(&:run)
@@ -91,10 +92,10 @@ class SyncChecker::RequestQueueTest < Minitest::Test
     result_set = []
     mutex = Mutex.new
     Typhoeus::Request.expects(:new).with(
-      "http://draft-content-store.dev.gov.uk/content/one",
+      "https://draft-content-store.test.alphagov.co.uk/content/one",
     ).returns(draft_request = stub(:on_complete))
     Typhoeus::Request.expects(:new).with(
-      "http://content-store.dev.gov.uk/content/two",
+      "https://content-store.test.alphagov.co.uk/content/two",
     ).returns(live_request = stub(:on_complete))
 
     queued_request = SyncChecker::RequestQueue.new(document_check, result_set, mutex)
